Change interface of XMPP tasks
The current examplary implementation handed over the UI element to be
manipulated upon completion of the task via the constructor, while the service
was handed over as a parameter to `execute()`, being passed into
`doInBackground()`.

This commit makes the service a parameter of the constructor as well, the policy
being that arguments to the method called on the service should be handed over
to `execute()` (probably being a `List<String>`) while objects that will
probably be required by every task (the `XMPPService` and a UI element to be
manipulated) should be provided in the constructor.