Add `geoloc_tester.py`
This commit introduces a very raw draft of a command line tool that dumps XMPP
messages received from a registered MUC and sends mock-up geoloc messages.

This simplifies debugging the process of sending and receiving geoloc messages
in the `TeamMeet` app and eliminates the need for a second instance of the app.