Fix `NullPointerException`
The `XmlPullParser` in the `GeolocExtensionProvider` iterates over every single
object of the XML tree as opposed to over every XML tag.

In order to get the text of a tag, the parser has to be advanced to the
`next()` object (the `TEXT` object) in the tree before calling `getText()`.
Calling `getText()` on a `START_TAG` will return `null`.

As a shortcut one may call `nextTag()` which advances the parser and returns the
text of the next object in one step.