Include groups in contacts
So far the result of `getContacts()` was a plain list of contact names. In this
data structure the information on which group a contact belongs to was lost.

This commit introduces a `Map<String, List<String>>` as the return type of
`getContacts()`. The keys are the groups of the roster and the values are lists
of user names contained in that group.