`getRoster()` from `XMPPService`
So far `getContacts()` was used to get a predigested list of contacts from the
roster.

Since the roster provides way more information than contact strings (e.g status
information, the ability to manage groups and contacts, ...), this commit
makes the `XMPPService` provide the raw roster which the `RosterActivity` uses
to fill an expandable list.

This eliminates the intermediate step of parsing the group and contact strings
from the roster that later need to be reworked into the very specific data
structures required by the `ExpandableListAdapter`.