Introduce `RosterListener`
Changes in presence are recognized. It is possible to find out which contact
sent the presence package and into which group that contact belongs.

It may be unfeasible though to retrieve the relative position of its group and
of itself within that group, which would be necessary to access the
corresponding child in the `ExpandableListAdapter` in order to update it.

It might be easier to rebuild the whole expandable list if a change in the
roster is detected.