Update random entry in `ExpandableListAdapter`
If we calculate the positions of the updated entry in the adapter, it is
possible to update the value. Although with the current code the list turns
unresponsive and redraws are only issued when triggered by pressing the menu
button.

Currently I tend to just rebuild the whole expandable list.