Introduce callbacks for `AsyncTask`s
So far `AsyncTask`s manipulated the UI by modifying references to `View`s they
got handed over in the constructor.

This commit introduces the `AsyncTaskCallback` interface that allows to pass an
object (potentially of an inner class) to an `AsyncTask` that will be called
upon completion and that will then handle UI updates.