Add default `AsyncTaskCallback`
So far all callback classes implemented the interface. Since the interface is
going to be extended to incorporate methods handling errors, the callbacks now
extend `BaseAsyncTaskCallback` which contains default implementations for the
methods required by the interface.

This will not only enable an incremental adoption of the extended interface but
more importantly allow callbacks registered to `AsyncTasks` that don't raise
errors to not implement empty, unused method stubs.