Move XMPP service back into activity
When transforming `RosterActivity` into `ContactsFragment`, it kept managing the
XMPP service.

Since both fragments will use it, this commit moves handling of the service back
into the overarching activity. The fragments are currently able to retrieve the
raw service via a getter from the activity.

Holding the service enables the activity to directly handle incoming intents
without having to refer them to a fragment and to handle the options menu by
itself.