Create new fragments in `RosterAdapter::getItem()`
So far the `RosterAdapter` was initialized with a list of fragments, created
everytime the `TabbedRosterActivity` was started. `getItem(pos)` returned the
item at the given position in that list.

This commit makes `getItem()` return a new instance of a fragment, where its
type is defined by the requested position. This is more in line with the usual
behaviour and use of `getItem()` and prevents the `TabbedRosterActivity` to have
to create fragments that may never be used (since existing ones are restored).