Get occupants of room
So far the `TeamsFragment` tried to fetch the occupants with role `Participant`
from each room when building its data structure.

Since this requires the role `Moderator`, this commit limits itself to fetching
the nick names of the occupants in each room.

NOTE: The returned list of occupants contains the user itself. In order to
filter oneself from the list, one needs to know one's nick in the room, which
might have been modified by the room to avoid collisions.