Filter self from occupants
So far the `TeamsFragment` included the nick of the user in the list of
occupants in each team.

This commit filters the user's nick from the list of occupants.

NOTE: Since we cannot filter without iterating, and we don't want to iterate
over each occupants list twice, we have to filter in the fragment before
displaying each occupant.
In order to know what to filter for, we might use the username from the
settings, provide access to the nickname of the/any room via the XMPP service or
provide access to the MUC itself.
This commit implements the second approach.