Broadcast team update event upon join
NOTE: We might want to introduce an async `JoinTeam` task. Else the joining will
happen in the `onClickListener` of the `Join` button in the `Group Invitation`
dialog, which freezes the UI until the action has been completed (which might be
a few seconds even on Wifi if the request fails).