Add `Team` wrapper
Wrap MUCs in class that in the future may hold associated information like the
session key used to encrypt messages send through the MUC and a list of invitees
that need receive the session key upon join.

NOTE: Extraction of common functionality will probably have to be performed when
we intend to add a generic MUC feature.