Track invitations
Add an `invitee` to a `Team` upon invitation. Remove (and handle) upon
decline/join.

NOTE: Upon join we only get the name portion of user, so we may only track
potentially non-matching nicks in the `Team`s.
NOTE: The listeners are reportedly unreliable and might have to be replaced with
raw packet listeners.