Handle crypto packets in `PacketListener`
So far `CryptoPacket`s were handled in a `MessageListener`. This added the
requirement for them to have a body and intermingled them with regular chat
messages.

This commit handles `CryptoPacket`s in a dedicated `PacketListener`. Since this
mechanism is more low level, it requires manual tracking of the `Team` the
transmitted keys are related to.