Add `ActionBar` logo
So far the `ActionBar` used the laucher icon as the app icon.

This commit introduces a dedicated logo that is used as the app icon in the
`ActionBar`. In contrast to the launcher icons, `ActionBar` logos don't have to
be square and may take advantage of the available horizontal space.