Don't display title in settings
So far the `ActionBar` in the `SettingsActivity` displayed the app logo as well
as the app title.

This commit disables displaying of the app name and limits itself to displaying
the launcher icon. This is not so much done for the need of screen real estate
but rather to provide a more consistent UI.