Fix aspect ratio of high density launcher icons
So far the launcher icons for high and extra high density displays were the same
as the `ActionBar` logos (i.e cropped to fit vs square).

This was potentially due to exporting the selection vs the page to bitmap in
inkscape.

This commit makes the high density launcher icons square again.