Remove chat drawer from map
So far the map contained a sliding drawer that demoed one possiblity to
integrate the related group chat into the map activity.

Since the drawer approach has been dropped (the chat is destined to become a tab
as soon as we managed to put the map in a fragment), this commit removes the
sliding drawer.