Encrypt and send session key
So far it was not possible to encrypt the session key.

This commit introduces the transformation from the Diffie-Hellman derived shared
secret into a 256 bit AES key, the encryption and consecutive dispatch of the
session key as well as the detection on the invitee end.

NOTE: There's no means yet to transmit the IV, so the invitee won't be able to
decrypt the AES session key.