Turn `ToastHelper` into singleton
So far each class that wanted to use a `ToastHelper` created a new instance of
it and passed along a reference to a `Context`.

Since not all classes have access to a `Context`, this commit has the
`XMPPService` initialize the `ToastHelper` upon creation. Each caller then
retrieves an initialized instance instead of creating a new one.