Add `password protected` field to form
So far when creating a team, the field indicating that the new MUC should be
password protected was always set to true. This resulted in an error if the form
sent by the server didn't contain that field (although standardized in XEP-0045
[0]).

This commit adds the field if missing in the form sent by the server.

 [0]: http://xmpp.org/extensions/xep-0045.html#registrar-formtype-owner