Extend `BaseExpandableListAdapter`
So far the Adapter filling the contact list was a `SimpleExpandableListAdapter`
which had two of its methods overriden.

This commit introduces a new custom adapter that extends
`BaseExpandableListAdapter` and uses two completely custom data structures to
represent its values.

NOTE: At this point, the data type representing a contact list child as well as
the corresponding layout are non-functional dummies.