Parse presence information
So far, the contact list only displayed a contact's Jabber ID and the
uninterpreted presence information.

This commit introduces the presentation of a contact's alias (iff set). In
addition, only contacts that are online are being shown (not configurable,
yet). The status message will display the mode and an optional status message
(iff set).