Get contact directly from adapter
So far the `onChildClick` handler of the contact list inflated a child layout
and retrieved a view in order to get to the name of the child that had been
selected.

This commit makes use of the new custom adapter and shortcuts above procedure by
directly asking the adapter for the name of the child at the given position.