Fix disabled connect button
So far the decision whether the connect button would be enabled or disabled was
made only when the options menu in the roster was refreshed. This led to a
disabled connect button until the next manual refresh in case that the XMPP
service wasn't available yet at the time the menu was constructed.

This commit automatically refreshes the options menu (and thereby enables the
`connect` button) when the binding to the XMPP service is complete. This avoids
that the user has to do so manually until the button becomes available.