Make ABS use patched support library
So far the ActionBarSherlock used google's original support library.

Since that one does not support maps as fragments, this commit removes the
original library and replaces it with the patched one by [pete doyle][0].

As the new library depends on `MapActivity`, the build target has to be set to
`Google APIs` to include the proprietary module.

 [0]: https://github.com/petedoyle/android-support-v4-googlemaps