Remove `ViewPager` from `TeamActivity`
So far the tabs in the team activity were managed by a `ViewPager` that allowed
for a horizontal swipe motion to change between tabs.

Since this interfered with the motions used to navigate within the map view,
this commit removes the view pager and controls the tabs exclusively via the
action bar.