Fix NPE when handling submenu item
So far the method handling context menu clicks tried to determine whether a
group or a contact in the expandable list had been clicked. This failed due to
the required information missing for submenu items (cf
https://github.com/nookieman/TeamMeet/commit/0cad6f0a65c6d665d5ed2d8783b6ed29aa929f5a#commitcomment-1917434).

This commit removes this check and indiscriminately switches over the ID of the
clicked menu item.