Fix refreshing of missed chat messages
So far the chat list in each `ChatFragment` was only refreshed `onCreateView()`
which led to new messages not being displayed until the fragment was destroyed
and recreated.

This commit puts the population of the chat list into `onResume()`, which will
be called when the fragment is shown after wasn't in focus (the message was
missed).

Fixes #192.